requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
aredis==1.1.8
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
bbcode==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.4.2
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.2
multidict==6.0.4
netaddr==0.9.0
orjson==3.9.10
Pillow==10.1.0
pycares==4.4.0
pycparser==2.21
pydantic==1.10.13
pymongo==4.6.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
rfc3339==6.2
setuptools==69.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
-e git+https://gitlab.gflclan.com/xy/xql.git@3dca00b346c9e0cdd3f638acea479c5ab966db66#egg=XQL
yarl==1.9.4
